+++ packages/mg-components/src/components/molecules/inputs/mg-input-checkbox/doc/mg-input-checkbox.stories.tsx
@@ -57,4 +57,5 @@ MgInputCheckbox.args = {
   tooltip: 'This is a tooltip',
   // Help Text
   helpText: 'Help text with html <strong>bold</strong>, <em>italic</em>.',
+  displaySelectedValues: false,
 };